app/requirements.txt
boto3
django
django-storages
psycopg2
pillow
uwsgi
libsass
django-compressor
django-sass-processor
django-allauth
